- Updated to latest orx version (string-id) - Updated windows binaries

+++ include/orx/base/orxDecl.h
@@ -60,6 +60,8 @@
 
 #endif /* TARGET_OS_ANDROID */
 
+#include <stddef.h>
+
 
 /* *** Platform dependent base declarations */
 
@@ -365,7 +367,7 @@
 
 
 /** Memory alignment macros */
-#define orxALIGN(ADDRESS, BLOCK_SIZE)   (((size_t)(ADDRESS) + ((BLOCK_SIZE) - 1)) & (~((BLOCK_SIZE) - 1)))
+#define orxALIGN(ADDRESS, BLOCK_SIZE)   (((size_t)(ADDRESS) + ((size_t)(BLOCK_SIZE) - 1)) & (~((size_t)(BLOCK_SIZE) - 1)))
 
 #define orxALIGN16(ADDRESS)             orxALIGN(ADDRESS, 16)
 #define orxALIGN32(ADDRESS)             orxALIGN(ADDRESS, 32)
